+++ noxfile.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
 
-import os
 import re
-import subprocess
 from pathlib import Path
-from typing import List, Tuple
 
 import nox
 from nox.sessions import Session