+++ src/django_idom/__init__.py
@@ -1,4 +1,4 @@
-from django_idom import components, decorators, hooks, types
+from django_idom import components, decorators, hooks, types, utils
 from django_idom.types import IdomWebsocket
 from django_idom.websocket.paths import IDOM_WEBSOCKET_PATH
 
@@ -6,9 +6,10 @@ from django_idom.websocket.paths import IDOM_WEBSOCKET_PATH
 __version__ = "1.1.0"
 __all__ = [
     "IDOM_WEBSOCKET_PATH",
-    "types",
     "IdomWebsocket",
     "hooks",
     "components",
     "decorators",
+    "types",
+    "utils",
 ]