+++ src/django_idom/config.py
@@ -1,11 +1,14 @@
 from typing import Dict
 
 from django.conf import settings
-from django.core.cache import DEFAULT_CACHE_ALIAS, caches
+from django.core.cache import DEFAULT_CACHE_ALIAS, BaseCache, caches
 from idom.core.types import ComponentConstructor
+
+from django_idom.types import ViewComponentIframe
 
 
 IDOM_REGISTERED_COMPONENTS: Dict[str, ComponentConstructor] = {}
+IDOM_VIEW_COMPONENT_IFRAMES: Dict[str, ViewComponentIframe] = {}
 
 IDOM_WEBSOCKET_URL = getattr(settings, "IDOM_WEBSOCKET_URL", "idom/")
 IDOM_WS_MAX_RECONNECT_TIMEOUT = getattr(
@@ -13,7 +16,8 @@ IDOM_WS_MAX_RECONNECT_TIMEOUT = getattr(
 )
 
 # Determine if using Django caching or LRU cache
-if "idom" in getattr(settings, "CACHES", {}):
-    IDOM_CACHE = caches["idom"]
-else:
-    IDOM_CACHE = caches[DEFAULT_CACHE_ALIAS]
+IDOM_CACHE: BaseCache = (
+    caches["idom"]
+    if "idom" in getattr(settings, "CACHES", {})
+    else caches[DEFAULT_CACHE_ALIAS]
+)