+++ src/django_idom/http/urls.py
@@ -10,5 +10,10 @@ urlpatterns = [
         "web_module/<path:file>",
         views.web_modules_file,  # type: ignore[arg-type]
         name="web_modules",
-    )
+    ),
+    path(
+        "iframe/<str:view_path>",
+        views.view_to_component_iframe,  # type: ignore[arg-type]
+        name="view_to_component",
+    ),
 ]