+++ tests/test_app/tests/test_components.py
@@ -10,8 +10,18 @@ class TestIdomCapabilities(ChannelsLiveServerTestCase):
     @classmethod
     def setUpClass(cls):
         if sys.platform == "win32":
-            raise SkipTest("These tests are broken on Windows due to Selenium")
+            raise SkipTest("These tests are broken on Windows.")
+
+            # FIXME: The following lines will be needed once Django channels fixes Windows tests
+            # See: https://github.com/django/channels/issues/1207
+
+            # import asyncio
+            # asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
+
+        # FIXME: This is required otherwise the tests will throw a `SynchronousOnlyOperation`
+        # error when deleting the test datatabase. Potentially a Django bug.
         os.environ["DJANGO_ALLOW_ASYNC_UNSAFE"] = "true"
+
         super().setUpClass()
         cls.playwright = sync_playwright().start()
         headed = bool(int(os.environ.get("PLAYWRIGHT_HEADED", 0)))
@@ -97,3 +107,72 @@ class TestIdomCapabilities(ChannelsLiveServerTestCase):
                 f"#todo-item-sample-{i}",
                 timeout=1,
             )
+
+    def test_view_to_component_sync_func(self):
+        self.page.locator("#view_to_component_sync_func[data-success=true]").wait_for()
+
+    def test_view_to_component_async_func(self):
+        self.page.locator("#view_to_component_async_func[data-success=true]").wait_for()
+
+    def test_view_to_component_sync_class(self):
+        self.page.locator("#ViewToComponentSyncClass[data-success=true]").wait_for()
+
+    def test_view_to_component_async_class(self):
+        self.page.locator("#ViewToComponentAsyncClass[data-success=true]").wait_for()
+
+    def test_view_to_component_template_view_class(self):
+        self.page.locator(
+            "#ViewToComponentTemplateViewClass[data-success=true]"
+        ).wait_for()
+
+    def _click_btn_and_check_success(self, name):
+        self.page.locator(f"#{name}:not([data-success=true])").wait_for()
+        self.page.wait_for_selector(f"#{name}_btn").click()
+        self.page.locator(f"#{name}[data-success=true]").wait_for()
+
+    def test_view_to_component_script(self):
+        self._click_btn_and_check_success("view_to_component_script")
+
+    def test_view_to_component_request(self):
+        self._click_btn_and_check_success("view_to_component_request")
+
+    def test_view_to_component_args(self):
+        self._click_btn_and_check_success("view_to_component_args")
+
+    def test_view_to_component_kwargs(self):
+        self._click_btn_and_check_success("view_to_component_kwargs")
+
+    def test_view_to_component_sync_func_compatibility(self):
+        self.page.frame_locator(
+            "#view_to_component_sync_func_compatibility > iframe"
+        ).locator(
+            "#view_to_component_sync_func_compatibility[data-success=true]"
+        ).wait_for()
+
+    def test_view_to_component_async_func_compatibility(self):
+        self.page.frame_locator(
+            "#view_to_component_async_func_compatibility > iframe"
+        ).locator(
+            "#view_to_component_async_func_compatibility[data-success=true]"
+        ).wait_for()
+
+    def test_view_to_component_sync_class_compatibility(self):
+        self.page.frame_locator(
+            "#view_to_component_sync_class_compatibility > iframe"
+        ).locator(
+            "#ViewToComponentSyncClassCompatibility[data-success=true]"
+        ).wait_for()
+
+    def test_view_to_component_async_class_compatibility(self):
+        self.page.frame_locator(
+            "#view_to_component_async_class_compatibility > iframe"
+        ).locator(
+            "#ViewToComponentAsyncClassCompatibility[data-success=true]"
+        ).wait_for()
+
+    def test_view_to_component_template_view_class_compatibility(self):
+        self.page.frame_locator(
+            "#view_to_component_template_view_class_compatibility > iframe"
+        ).locator(
+            "#ViewToComponentTemplateViewClassCompatibility[data-success=true]"
+        ).wait_for()